refactor: reorganize imports and remove RunnableConstant from basic module
