+++ tests/test_basic.py
@@ -1,25 +1,7 @@
 import pytest
-from runnable_family.standard import (
+from runnable_family.basic import (
     RunnableAdd,
-    RunnableConstant,
 )
-
-
-@pytest.mark.parametrize(
-    'input_obj, expected',
-    [
-        (0, 0),
-        ('a', 0),
-        ({'a': 1}, 0),
-        ({'a': 1}, 1),
-    ]
-)
-def test_runnable_constant(
-    input_obj,
-    expected,
-):
-    constant = RunnableConstant(expected)
-    assert constant.invoke(input_obj) == expected
 
 
 @pytest.mark.parametrize(
